(WEBRTC-519) Fix cannot create channel error
Once we have created the offer, the ICE candidates negotiation process start. This process can return none or several ICE candidates  (at least one ICE candidate is requried to start calling).
Mentioned that, we have implemented a timer that TIMEOUTS after 300 mSeconds. Each time an ICE candidate is received, we reset the timer. When there are no more ICE candidates, the timer will timeout and call the appropiated delgate to send the invite message.

There was an issue inside the timer logic producing multiple calls of the delegate, producing multiple invite messages been sent to the remote peer.

On this change we are avoiding the error by chekcing that the timer has timeouted before calling the delegate.